+++ src/main/java/uz/master/demotest/controller/teacher/TeacherController.java
@@ -5,8 +5,10 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import uz.master.demotest.dto.test.OverAllResultDTO;
 import uz.master.demotest.dto.test.TestCreateDto;
+import uz.master.demotest.entity.result.OverAllResult;
 import uz.master.demotest.entity.test.Subject;
 import uz.master.demotest.entity.test.Test;
+import uz.master.demotest.enums.Role;
 import uz.master.demotest.services.GroupService;
 import uz.master.demotest.services.SubjectService;
 import uz.master.demotest.services.auth.AuthUserService;
@@ -25,7 +27,6 @@ public class TeacherController {
     private final SessionUser sessionUser;
     private final SubjectService subjectService;
     private final GroupService groupService;
-    private final OverAllResultService overAllResultService;
 
     public TeacherController(AuthUserService authUserService, TestService testService, SessionUser sessionUser, SubjectService subjectService, GroupService groupService, OverAllResultService overAllResultService) {
         this.authUserService = authUserService;
@@ -33,11 +34,13 @@ public class TeacherController {
         this.sessionUser = sessionUser;
         this.subjectService = subjectService;
         this.groupService = groupService;
-        this.overAllResultService = overAllResultService;
     }
 
     @GetMapping(value = "myTests")
-    public String getResults(Model model) {
+    public String myTests(Model model) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
         model.addAttribute("user", sessionUser.getFullName());
         model.addAttribute("tests", testService.getMyOwnerTests());
         return "teacher/myTests";
@@ -45,6 +48,9 @@ public class TeacherController {
 
     @GetMapping(value = "/createTest")
     public String getCreatePage(Model model) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
         model.addAttribute("user", sessionUser.getFullName());
         List<Subject> subjects = subjectService.getAllSubjects();
         model.addAttribute("subjects", subjects);
@@ -53,6 +59,9 @@ public class TeacherController {
 
     @GetMapping(value = "/createTest?error")
     public String getCreatePageWithError(Model model) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
         model.addAttribute("error", "Bunday nomli test mavjud yoki ma'lumotlar to'liq emas");
         model.addAttribute("user", sessionUser.getFullName());
         List<Subject> subjects = subjectService.getAllSubjects();
@@ -62,6 +71,9 @@ public class TeacherController {
 
     @GetMapping(value = "/editTest/{id}")
     public String getEditPage(@PathVariable Long id, Model model) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
         Test testById = testService.getTest(id);
         model.addAttribute("test", testById);
         List<Subject> subjects = subjectService.getAllSubjects();
@@ -72,6 +84,9 @@ public class TeacherController {
 
     @PostMapping(value = "/editTest")
     public String updateTest(@ModelAttribute Test test) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
         try {
             testService.updateTest(test);
         } catch (Exception e){
@@ -82,24 +97,36 @@ public class TeacherController {
 
     @GetMapping(value = "/createQuestion")
     public String getCreateQuestionPage(Model model) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
         model.addAttribute("user", sessionUser.getFullName());
         return "/teacher/createQuestion";
     }
 
     @GetMapping(value = "/doAction/{id}")
     public String doAction(@PathVariable Long id) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
         testService.doAction(id);
         return "redirect:/teacher/myTests";
     }
 
     @GetMapping(value = "/deleteTest/{id}")
     public String delete(@PathVariable Long id) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
         testService.delete(id);
         return "redirect:/teacher/myTests";
     }
 
     @PostMapping(value = "/createTest")
     public String createTest(@ModelAttribute TestCreateDto dto) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
         try {
             Long testId = testService.createTest(dto);
             testService.restoreExcelFileToTest(testId, dto.getFile());
@@ -111,13 +138,14 @@ public class TeacherController {
     }
 
     @GetMapping(value = "/results")
-    public String getResult(Model model) {
+    public String getResults(Model model) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
         model.addAttribute("user", sessionUser.getFullName());
         try {
-            List<OverAllResultDTO> results = testService.getAllResultsForTeacher();
-            model.addAttribute("groups", groupService.getAllGroups());
-            model.addAttribute("results", results);
-
+            List<Test> tests = testService.getAllResultsForTeacher();
+            model.addAttribute("tests", tests);
         } catch (Exception ex) {
             model.addAttribute("error", "Ma'lumotlar topilmadi!");
             return "/error/error";
@@ -125,19 +153,45 @@ public class TeacherController {
         return "/teacher/results";
     }
 
-    @GetMapping(value = "/results/{groupName}")
-    public String getAllStudentsByGroup(Model model, @PathVariable String groupName) {
-        Long id = authUserService.setAuthUserSelectedGroup(groupName);
-        List<OverAllResultDTO> allResult = testService.getAllResultsByGroup(id, groupName);
+    @GetMapping(value = "/result/{id}")
+    public String getResult(Model model, @PathVariable Long id) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
+        Long userId = authUserService.setAuthUserSelectedGroup(null);
+        List<OverAllResult> allResult = testService.getOverAllResultByTestId(id);
+        model.addAttribute("id", id);
         model.addAttribute("groups", groupService.getAllGroups());
         model.addAttribute("user", sessionUser.getFullName());
-        model.addAttribute("results", allResult);
-        return "/teacher/results";
+        model.addAttribute("allResults", allResult);
+        return "/teacher/result";
+    }
+
+    @GetMapping(value = "/result/{testId}/{groupName}")
+    public String getAllStudentsByGroup(Model model, @PathVariable String groupName, @PathVariable Long testId) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
+        try{
+            Long userId = authUserService.setAuthUserSelectedGroup(groupName);
+            List<OverAllResult> allResult = testService.getOverAllResultByTestIdAndGroupNameForTeacher(testId, groupName);
+            model.addAttribute("id", testId);
+            model.addAttribute("groups", groupService.getAllGroups());
+            model.addAttribute("user", sessionUser.getFullName());
+            model.addAttribute("allResults", allResult);
+        } catch(Exception e){
+            model.addAttribute("error", "Bu sizning testingiz emas yoki bunday testmavjud emas!!!");
+            return "/error/error";
+        }
+        return "/teacher/result";
     }
 
 
     @GetMapping(value = "/checkDeleteTest/{id}")
     public String checkDeleteTest(Model model, @PathVariable Long id) {
+        if(!authUserService.checkToRole(Role.TEACHER)){
+            return "redirect:/home";
+        }
         model.addAttribute("testName", testService.getTestById(id));
         model.addAttribute("testId", id);
         model.addAttribute("user", sessionUser.getFullName());
